refactor: #3 component 폴더 네이밍을 component 이름 형식과 일치하게 변경
