+++ memoria-web/src/components/EmotionItem/EmotionItem.css
