+++ memoria-web/src/components/EmotionItem/EmotionItem.jsx
